+++ src/main/java/com/h3/reservation/controller/BotController.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.h3.reservation.slack.RequestType;
-import com.h3.reservation.slack.dto.EventCallbackRequestDto;
-import com.h3.reservation.slack.dto.VerificationRequestDto;
+import com.h3.reservation.slack.dto.request.EventCallbackRequest;
+import com.h3.reservation.slack.dto.request.VerificationRequest;
 import com.h3.reservation.slack.service.SlackService;
 import net.minidev.json.JSONObject;
 import org.slf4j.Logger;
@@ -36,10 +36,10 @@ public class BotController {
     public ResponseEntity action(@RequestBody JSONObject req) throws JsonProcessingException {
         RequestType type = RequestType.valueOf(req.getAsString("type").toUpperCase());
         if (RequestType.URL_VERIFICATION.equals(type)) {
-            return ResponseEntity.ok(service.verify(toDto(req, VerificationRequestDto.class)));
+            return ResponseEntity.ok(service.verify(toDto(req, VerificationRequest.class)));
         }
         if (RequestType.EVENT_CALLBACK.equals(type)) {
-            service.eventCallBack(toDto(req, EventCallbackRequestDto.class));
+            service.eventCallBack(toDto(req, EventCallbackRequest.class));
             return ResponseEntity.ok().build();
         }
         return ResponseEntity.badRequest().build();
