+++ src/main/java/com/h3/reservation/slack/dto/request/VerificationRequest.java
@@ -1,19 +1,19 @@
-package com.h3.reservation.slack.dto;
+package com.h3.reservation.slack.dto.request;
 
 /**
  * @author heebg
  * @version 1.0
  * @date 2019-12-02
  */
-public class VerificationRequestDto {
+public class VerificationRequest {
     private String token;
     private String challenge;
     private String type;
 
-    public VerificationRequestDto() {
+    public VerificationRequest() {
     }
 
-    public VerificationRequestDto(String token, String challenge, String type) {
+    public VerificationRequest(String token, String challenge, String type) {
         this.token = token;
         this.challenge = challenge;
         this.type = type;
